fix(postgres): pgx simple-protocol type-encoding issues (#230)

+++ internal/adapters/storage/postgres/purchase_price_store.go
@@ -22,11 +22,14 @@ func (ps *PurchaseStore) UpdatePurchasePriceOverride(ctx context.Context, purcha
 	}
 	// Committing an override supersedes any outstanding AI suggestion — clear it
 	// so the item stops showing up in "Needs Attention" on that signal.
+	// $1 is reused in three contexts; pin type as BIGINT to avoid
+	// SQLSTATE 42P08 ("inconsistent types deduced for parameter") under
+	// pgx QueryExecModeExec. Same fix as pricing_store.SetReviewedPrice.
 	result, err := ps.db.ExecContext(ctx,
 		`UPDATE campaign_purchases
-		 SET override_price_cents = $1, override_source = $2, override_set_at = $3,
-		     ai_suggested_price_cents = CASE WHEN $1 > 0 THEN 0 ELSE ai_suggested_price_cents END,
-		     ai_suggested_at         = CASE WHEN $1 > 0 THEN '' ELSE ai_suggested_at END,
+		 SET override_price_cents = $1::BIGINT, override_source = $2, override_set_at = $3,
+		     ai_suggested_price_cents = CASE WHEN $1::BIGINT > 0 THEN 0 ELSE ai_suggested_price_cents END,
+		     ai_suggested_at         = CASE WHEN $1::BIGINT > 0 THEN '' ELSE ai_suggested_at END,
 		     updated_at = $4
 		 WHERE id = $5`,
 		priceCents, source, setAt, now, purchaseID,
